use int64 for Time and Elapsed fields in History struct

+++ history.go
@@ -2,7 +2,7 @@ package zshhist
 
 // History represents a command history
 type History struct {
-	Time    int
-	Elapsed int
+	Time    int64
+	Elapsed int64
 	Command string
 }
